Only cleanup pods, on the specific node, which fixes reboot pod was deleted before it was started. General code cleanup.
